# src/docs/swagger.yaml
definitions:
  api_handlers.personData:
    properties:
      first_name:
        maxLength: 10
        minLength: 4
        type: string
      last_name:
        maxLength: 20
        minLength: 6
        type: string
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
      - first_name
      - last_name
      - mobile_number
    type: object
...
  github_com_nazarow_golang-clean_api_dto.CreatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      unit:
        maxLength: 15
        type: string
    required:
      - categoryId
      - name
    type: object
...
  github_com_nazarow_golang-clean_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description:
          contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_PropertyResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyResponse"
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Filter"
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Sort"
        type: array
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PropertyCategoryResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyResponse"
        type: array
    type: object
  github_com_nazarow_golang-clean_api_dto.PropertyResponse:
    properties:
      category:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyCategoryResponse"
      dataType:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      unit:
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.UpdatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        type: string
      unit:
        maxLength: 15
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.ResultCode"
      success:
        type: boolean
      validationErrors:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_validations.ValidationError"
        type: array
    type: object
  github_com_nazarow_golang-clean_api_helper.ResultCode:
    enum:
      - 0
      - 40001
      - 40101
      - 40301
      - 40401
      - 42901
      - 42902
      - 50001
      - 50002
    type: integer
    x-enum-varnames:
      - Success
      - ValidationError
      - AuthError
      - ForbiddenError
      - NotFoundError
      - LimiterError
      - OtpLimiterError
      - CustomRecovery
      - InternalError
  github_com_nazarow_golang-clean_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/properties/:
    post:
      consumes:
        - application/json
      description: Create a Property
      parameters:
        - description: Create a Property
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CreatePropertyRequest"
      responses:
        "201":
          description: Property response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Create a Property
      tags:
        - Properties
  /v1/properties/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a Property
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Delete a Property
      tags:
        - Properties
    get:
      consumes:
        - application/json
      description: Get a Property
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: Property response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get a Property
      tags:
        - Properties
    put:
      consumes:
        - application/json
      description: Update a Property
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
        - description: Update a Property
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.UpdatePropertyRequest"
      responses:
        "200":
          description: Property response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PropertyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Update a Property
      tags:
        - Properties
  /v1/properties/get-by-filter:
    post:
      consumes:
        - application/json
      description: Get Properties
      parameters:
        - description: Request
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter"
      responses:
        "200":
          description: Property response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_PropertyResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get Properties
      tags:
        - Properties
...
  /v1/test/binder/body:
    post:
      consumes:
        - application/json
      description: BodyBinder
      parameters:
        - description: person data
          in: body
          name: person
          required: true
          schema:
            $ref: "#/definitions/api_handlers.personData"
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  validationErrors:
                    type: object
                type: object
        "400":
          description: Failed
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: BodyBinder
      tags:
        - Test
  /v1/test/binder/uri/{id}/{name}:
    post:
      consumes:
        - application/json
      description: BodyBinder
      parameters:
        - description: user id
          in: path
          name: id
          required: true
          type: integer
        - description: user name
          in: path
          name: name
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  validationErrors:
                    type: object
                type: object
        "400":
          description: Failed
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: BodyBinder
      tags:
        - Test
  /v1/test/user/{id}:
    get:
      consumes:
        - application/json
      description: UserById
      parameters:
        - description: user id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Failed
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      summary: UserById
      tags:
        - Test
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"